Revrt posix timer back to alarm

See merge request nirs/blkhash!113

+++ test/blkhash-bench.c
@@ -11,7 +11,7 @@
 
 static enum input_type input_type = DATA;
 static const char *digest_name = "sha256";
-static double timeout_seconds = 1.0;
+static int timeout_seconds = 1;
 static int64_t input_size = 0;
 static int block_size = 64 * KiB;
 static int read_size = 1 * MiB;
@@ -225,7 +225,7 @@ int main(int argc, char *argv[])
     printf("{\n");
     printf("  \"input-type\": \"%s\",\n", type_name(input_type));
     printf("  \"digest-name\": \"%s\",\n", digest_name);
-    printf("  \"timeout-seconds\": %.3f,\n", timeout_seconds);
+    printf("  \"timeout-seconds\": %d,\n", timeout_seconds);
     printf("  \"input-size\": %" PRIi64 ",\n", input_size);
     printf("  \"block-size\": %d,\n", block_size);
     printf("  \"read-size\": %d,\n", read_size);
